refactor IrcClient
This mostly splits off different parts of IrcClient into different files.
Note that even this is only a temporary "solution" (not necessarily
implying there's a problem with the current code base). If you wanted to
implement this even "cleaner", you'd probably have to split the IrcClient
class into several seperate classes altogether. As this would however
result in a whole bunch of public API changes, I've refrained from it in
this PR.
NOTE: this also fixes several small bugs, mostly related to copy/paste.
I haven't had a lot of time to look over the rest of the code base, but
maybe other refactoring commits will follow. I just thought I'd start with
this because it seems to be the biggest and most complex mess.